Move `connector` config to the root level
