+++ sqrl-planner/src/main/resources/default-package.json
@@ -15,68 +15,6 @@
         "state.backend.type": "rocksdb",
         "table.exec.resource.default-parallelism": 1,
         "taskmanager.memory.network.max": "800m"
-      },
-      "connectors": {
-        "postgres": {
-          "connector": "jdbc-sqrl",
-          "password": "${JDBC_PASSWORD}",
-          "driver": "org.postgresql.Driver",
-          "username": "${JDBC_USERNAME}",
-          "url": "jdbc:postgresql://${JDBC_AUTHORITY}",
-          "table-name": "${sqrl:table-name}"
-        },
-        "kafka-mutation": {
-          "connector" : "kafka",
-          "format" : "flexible-json",
-          "flexible-json.timestamp-format.standard": "ISO-8601",
-          "properties.bootstrap.servers": "${PROPERTIES_BOOTSTRAP_SERVERS}",
-          "properties.group.id": "${PROPERTIES_GROUP_ID}",
-          "properties.auto.offset.reset" : "earliest",
-          "topic" : "${sqrl:table-name}"
-        },
-        "kafka": {
-          "connector" : "kafka",
-          "format" : "flexible-json",
-          "flexible-json.timestamp-format.standard": "ISO-8601",
-          "properties.bootstrap.servers": "${PROPERTIES_BOOTSTRAP_SERVERS}",
-          "properties.group.id": "${PROPERTIES_GROUP_ID}",
-          "topic" : "${sqrl:table-name}"
-        },
-        "localfile": {
-          "connector" : "filesystem",
-          "format" : "${sqrl:format}",
-          "path" : "${DATA_PATH}/${sqrl:filename}",
-          "source.monitor-interval" : "10 sec"
-        },
-        "iceberg": {
-          "connector": "iceberg",
-          "catalog-table": "${sqrl:table-name}"
-        },
-        "postgres_log-source": {
-          "connector": "postgres-cdc",
-          "hostname": "${PGHOST}",
-          "port": "5432",
-          "username": "${JDBC_USERNAME}",
-          "password": "${JDBC_PASSWORD}",
-          "database-name": "datasqrl",
-          "schema-name": "public",
-          "table-name": "${sqrl:table-name}",
-          "slot.name": "flink_slot",
-          "decoding.plugin.name": "pgoutput",
-          "debezium.slot.drop_on_stop": "false"
-        },
-        "postgres_log-sink": {
-          "connector": "jdbc-sqrl",
-          "password": "${JDBC_PASSWORD}",
-          "driver": "org.postgresql.Driver",
-          "username": "${JDBC_USERNAME}",
-          "url": "jdbc:postgresql://${JDBC_AUTHORITY}",
-          "table-name": "${sqrl:table-name}"
-        },
-        "print": {
-          "connector": "print",
-          "print-identifier": "${sqrl:table-name}"
-        }
       }
     },
     "snowflake" : {
@@ -84,6 +22,68 @@
       "catalog-name": "${SNOWFLAKE_CATALOG_NAME}",
       "external-volume": "${SNOWFLAKE_EXTERNAL_VOLUME}",
       "url": "jdbc:snowflake://${SNOWFLAKE_ID}.snowflakecomputing.com/?user=${SNOWFLAKE_USER}&password=${SNOWFLAKE_PASSWORD}&warehouse=COMPUTE_WH&db=MYSNOWFLAKEDB&schema=public&disableSslHostnameVerification=true"
+    }
+  },
+  "connectors": {
+    "postgres": {
+      "connector": "jdbc-sqrl",
+      "password": "${JDBC_PASSWORD}",
+      "driver": "org.postgresql.Driver",
+      "username": "${JDBC_USERNAME}",
+      "url": "jdbc:postgresql://${JDBC_AUTHORITY}",
+      "table-name": "${sqrl:table-name}"
+    },
+    "kafka-mutation": {
+      "connector" : "kafka",
+      "format" : "flexible-json",
+      "flexible-json.timestamp-format.standard": "ISO-8601",
+      "properties.bootstrap.servers": "${PROPERTIES_BOOTSTRAP_SERVERS}",
+      "properties.group.id": "${PROPERTIES_GROUP_ID}",
+      "properties.auto.offset.reset": "earliest",
+      "topic" : "${sqrl:table-name}"
+    },
+    "kafka": {
+      "connector" : "kafka",
+      "format" : "flexible-json",
+      "flexible-json.timestamp-format.standard": "ISO-8601",
+      "properties.bootstrap.servers": "${PROPERTIES_BOOTSTRAP_SERVERS}",
+      "properties.group.id": "${PROPERTIES_GROUP_ID}",
+      "topic" : "${sqrl:table-name}"
+    },
+    "localfile": {
+      "connector": "filesystem",
+      "format": "${sqrl:format}",
+      "path": "${DATA_PATH}/${sqrl:filename}",
+      "source.monitor-interval" : "10 sec"
+    },
+    "iceberg": {
+      "connector": "iceberg",
+      "catalog-table": "${sqrl:table-name}"
+    },
+    "postgres_log-source": {
+      "connector": "postgres-cdc",
+      "hostname": "${PGHOST}",
+      "port": "5432",
+      "username": "${JDBC_USERNAME}",
+      "password": "${JDBC_PASSWORD}",
+      "database-name": "datasqrl",
+      "schema-name": "public",
+      "table-name": "${sqrl:table-name}",
+      "slot.name": "flink_slot",
+      "decoding.plugin.name": "pgoutput",
+      "debezium.slot.drop_on_stop": "false"
+    },
+    "postgres_log-sink": {
+      "connector": "jdbc-sqrl",
+      "password": "${JDBC_PASSWORD}",
+      "driver": "org.postgresql.Driver",
+      "username": "${JDBC_USERNAME}",
+      "url": "jdbc:postgresql://${JDBC_AUTHORITY}",
+      "table-name": "${sqrl:table-name}"
+    },
+    "print": {
+      "connector": "print",
+      "print-identifier": "${sqrl:table-name}"
     }
   },
   "test-runner": {
