+++ sqrl-planner/src/test/resources/config/package/emptyEnginesFlinkConnectors.json
@@ -1,7 +1,3 @@
 {
-  "engines": {
-    "flink": {
-      "connectors": {}
-    }
-  }
+  "connectors": {}
 }
