+++ sqrl-testing/sqrl-integration-tests/src/test/resources/dagplanner/package.json
@@ -1,18 +1,14 @@
 {
   "version": "1",
-  "engines" : {
-    "flink" : {
-      "connectors": {
-        "kafka": {
-          "topic" : "kafka-${sqrl:table-name}"
-        },
-        "kafka-mutation": {
-          "topic" : "kafka-mutation-${sqrl:table-name}"
-        },
-        "postgres": {
-          "table-name": "${sqrl:table-id}"
-        }
-      }
+  "connectors": {
+    "kafka": {
+      "topic" : "kafka-${sqrl:table-name}"
+    },
+    "kafka-mutation": {
+      "topic" : "kafka-mutation-${sqrl:table-name}"
+    },
+    "postgres": {
+      "table-name": "${sqrl:table-id}"
     }
   }
 }
