+++ sqrl-testing/sqrl-integration-tests/src/test/resources/usecases/analytics-only/package-analytics-only.json
@@ -12,15 +12,15 @@
         "table.exec.source.idle-timeout": "1 ms",
         "execution.checkpointing.interval": "6 s",
         "execution.checkpointing.min-pause": "2 s"
-      },
-      "connectors" : {
-        "iceberg" : {
-          "write.parquet.page-size-bytes": 1000,
-          "warehouse":"/tmp/duckdb",
-          "catalog-type":"hadoop",
-          "catalog-name": "mydatabase"
-        }
       }
+    }
+  },
+  "connectors" : {
+    "iceberg" : {
+      "write.parquet.page-size-bytes": 1000,
+      "warehouse":"/tmp/duckdb",
+      "catalog-type":"hadoop",
+      "catalog-name": "mydatabase"
     }
   },
   "test-runner": {
