+++ sqrl-testing/sqrl-integration-tests/src/test/resources/usecases/analytics-only/package-snowflake.json
@@ -9,15 +9,15 @@
     "flink": {
       "config": {
         "table.exec.source.idle-timeout": "1 ms"
-      },
-      "connectors" : {
-        "iceberg" : {
-          "write.parquet.page-size-bytes": 1000,
-          "warehouse":"/tmp/duckdb",
-          "catalog-type":"hadoop",
-          "catalog-name": "mydatabase"
-        }
       }
+    }
+  },
+  "connectors" : {
+    "iceberg" : {
+      "write.parquet.page-size-bytes": 1000,
+      "warehouse":"/tmp/duckdb",
+      "catalog-type":"hadoop",
+      "catalog-name": "mydatabase"
     }
   },
   "test-runner": {
